requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator>=1.3.0
requests>=2.25.0
python-dotenv>=0.19.0

# NLP and AI Dependencies
transformers>=4.21.0
torch>=1.12.0
nltk>=3.7
scikit-learn>=1.1.0
numpy>=1.21.0
